[Feat] 생산실적 공장별 월별 6개월 누적 생산량 조회 구현 완료.
