+++ CtrlLine/src/main/java/com/beyond/synclab/ctrlline/domain/productionperformance/repository/query/ProductionPerformanceAllQueryRepository.java
@@ -1,4 +1,4 @@
-package com.beyond.synclab.ctrlline.domain.productionperformance.repository.query.all;
+package com.beyond.synclab.ctrlline.domain.productionperformance.repository.query;
 
 import com.beyond.synclab.ctrlline.domain.productionperformance.dto.request.SearchAllProductionPerformanceRequestDto;
 import com.beyond.synclab.ctrlline.domain.productionperformance.dto.response.GetAllProductionPerformanceResponseDto;
