+++ CtrlLine/src/main/java/com/beyond/synclab/ctrlline/domain/productionperformance/repository/query/ProductionPerformanceMonthlyDefRateQueryRepositoryImpl.java
@@ -1,19 +1,16 @@
-package com.beyond.synclab.ctrlline.domain.productionperformance.repository.query.monthlysum;
+package com.beyond.synclab.ctrlline.domain.productionperformance.repository.query;
 
 import com.beyond.synclab.ctrlline.domain.itemline.entity.QItemsLines;
 import com.beyond.synclab.ctrlline.domain.line.entity.QLines;
 import com.beyond.synclab.ctrlline.domain.productionperformance.entity.QProductionPerformances;
 import com.beyond.synclab.ctrlline.domain.productionplan.entity.QProductionPlans;
-
 import com.querydsl.core.Tuple;
 import com.querydsl.core.types.dsl.Expressions;
 import com.querydsl.core.types.dsl.NumberExpression;
 import com.querydsl.core.types.dsl.StringExpression;
 import com.querydsl.jpa.impl.JPAQueryFactory;
-
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-
 import org.springframework.stereotype.Repository;
 
 import java.math.BigDecimal;
@@ -25,39 +22,42 @@ import java.util.Map;
 @Slf4j
 @Repository
 @RequiredArgsConstructor
-public class ProductionPerformanceMonthlyQueryRepositoryImpl
-        implements ProductionPerformanceMonthlyQueryRepository {
+public class ProductionPerformanceMonthlyDefRateQueryRepositoryImpl
+        implements ProductionPerformanceMonthlyDefRateQueryRepository {
 
     private final JPAQueryFactory queryFactory;
 
     @Override
-    public Map<YearMonth, Long> getMonthlySum(String factoryCode, List<YearMonth> months) {
+    public Map<YearMonth, MonthlyQtySum> getMonthlyQtySum(String factoryCode, List<YearMonth> months) {
 
         QProductionPerformances perf = QProductionPerformances.productionPerformances;
         QProductionPlans plan = QProductionPlans.productionPlans;
         QItemsLines itemLine = QItemsLines.itemsLines;
         QLines line = QLines.lines;
 
-        // "YYYY-MM" 리스트
+        // "YYYY-MM" 문자열 리스트
         List<String> monthStrings = months.stream()
                 .map(YearMonth::toString)
                 .toList();
 
-        // YYYY/MM → YYYY-MM
-        // substring + replace 적용
+        // YYYY/MM → YYYY-MM (전표번호에서 월 추출)
         StringExpression ymExpr = Expressions.stringTemplate(
                 "replace(substring({0}, 1, 7), '/', '-')",
                 perf.performanceDocumentNo
         );
 
-        // DB에서 BigDecimal SUM → 이후 Java에서 Long 변환
-        NumberExpression<BigDecimal> sumExpr =
+        // SUM(totalQty), SUM(performanceQty) (BigDecimal → Long 변환 예정)
+        NumberExpression<BigDecimal> totalQtySumExpr =
+                Expressions.numberTemplate(BigDecimal.class, "sum({0})", perf.totalQty);
+
+        NumberExpression<BigDecimal> performanceQtySumExpr =
                 Expressions.numberTemplate(BigDecimal.class, "sum({0})", perf.performanceQty);
 
         List<Tuple> result = queryFactory
                 .select(
                         ymExpr.as("month"),
-                        sumExpr
+                        totalQtySumExpr,
+                        performanceQtySumExpr
                 )
                 .from(perf)
                 .leftJoin(perf.productionPlan, plan)
@@ -70,18 +70,23 @@ public class ProductionPerformanceMonthlyQueryRepositoryImpl
                 .groupBy(ymExpr)
                 .fetch();
 
-        Map<YearMonth, Long> sumByMonth = new HashMap<>();
+        Map<YearMonth, MonthlyQtySum> sumByMonth = new HashMap<>();
 
         for (Tuple tuple : result) {
 
             String ym = tuple.get(0, String.class);
 
-            // Long 변환 방식
-            BigDecimal totalDecimal = tuple.get(1, BigDecimal.class);
-            Long total = (totalDecimal != null) ? totalDecimal.longValue() : 0L;
+            BigDecimal totalQtyDecimal = tuple.get(1, BigDecimal.class);
+            BigDecimal performanceQtyDecimal = tuple.get(2, BigDecimal.class);
+
+            BigDecimal totalQtySum = totalQtyDecimal != null ? totalQtyDecimal : BigDecimal.ZERO;
+            BigDecimal performanceQtySum = performanceQtyDecimal != null ? performanceQtyDecimal : BigDecimal.ZERO;
 
             if (ym != null) {
-                sumByMonth.put(YearMonth.parse(ym), total);
+                sumByMonth.put(
+                        YearMonth.parse(ym),
+                        new MonthlyQtySum(totalQtySum, performanceQtySum)
+                );
             }
         }
